Update conf.py
License CC-BY-SA 4.0 added. As this is a derivated work from Blender manual (license under CC-BY-SA 4.0) and we will try to use the "Game Engine Book" by Mike Pan and Dalai Felinto (licensed also under CC-BY-SA 4.0) we have to license this manual under the same license.
Later, I will upload a new "About this manual" chapter, similar to Blender manual one, clarifying the different authors as it is required in the CC-BY-SA 4.0 license.